Rename namespace from Dashboard to Home
