+++ osu.Game/Overlays/Dashboard/Home/DashboardPopularBeatmapPanel.cs
@@ -9,7 +9,7 @@ using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
 using osuTK;
 
-namespace osu.Game.Overlays.Dashboard.Dashboard
+namespace osu.Game.Overlays.Dashboard.Home
 {
     public class DashboardPopularBeatmapPanel : DashboardBeatmapPanel
     {
