+++ osu.Game/Overlays/Dashboard/Home/DrawableBeatmapList.cs
@@ -11,7 +11,7 @@ using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
 using osuTK;
 
-namespace osu.Game.Overlays.Dashboard.Dashboard
+namespace osu.Game.Overlays.Dashboard.Home
 {
     public abstract class DrawableBeatmapList : CompositeDrawable
     {
